Major changes to the project (daemon)
The way keys are handled has been modified. Now, there is a REST server that manages public keys and logged-in users. This server integrates with the DB and publishes the public keys of the users. The server presents an initial implementation of threshold sharing using Shamir. In a future version, users will be able to request their own Shamir keys (encrypted with their public key) from the server, and then reassemble them on the client.