Implement key handling with keyutils
Key management is hangled by key_utils library that is a simple facede for the Keyutils Linux system library. This allows us to put a key in the kernel and fetch it when needed. The password option in the configuration file has been removed as it is not necessary anymore