Update the structure of the crypt-utils library and implemet CTR encryption
The crypt-utils library has been updated. It is now a standalone library that contains all the encryption function. A makefile for the library was added and the main makefile was modified to accomodate this change. Two new functions were written to use AES CTR encryption/decryption. The plan is to make this functions use more than one thread (future updates).